Initial refactor of ssr into setup-server
