+++ packages/extension-chakra-storefront/src/middleware/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2025, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import jwksCachingMiddleware from './jwks-caching'
+import passwordResetMiddleware from './password-reset'
+import passwordlessMiddleware from './passwordless'
+
+export {jwksCachingMiddleware, passwordlessMiddleware, passwordResetMiddleware}
